Units now have a defense value.

+++ src/text_input.c
@@ -54,23 +54,12 @@ void text_input_update(TextInput *text_input) {
 void text_input_draw(TextInput *text_input) {
 
 	int size = MeasureText(text_input->target, 10);
+
 	Vector2 pos = {
-   		floor((480 / 2.0f) - (size / 2)),
-		floor((270 / 3.0f) - (10 / 2)),
+   		floor((480 / 2.0f) - ((float)size / 2)),
+		floor((270 / 3.0f) - ((float)10 / 2)),
 	};
 
-	DrawText(
-		text_input->target,
-		pos.x,
-		pos.y,
-		10,
-		WHITE
-	);
-	DrawText(
-		text_input->buff,
-		pos.x,
-		pos.y,
-		10,
-		DARKPURPLE
-	);
+	DrawText(text_input->target, pos.x, pos.y, 10, WHITE);
+	DrawText( text_input->buff, pos.x, pos.y, 10, DARKPURPLE);
 }
